// src/app/models/acts.model.ts
export class Acts{
    fechaSus: Date;
    autor: string;
    codigoEstudiante: number;
    programa: string;
    nombreProyecto: string;
    tutor: string;
    notaTutor: number;
    jurado1: string;
    notaJurado1: number;
    jurado2: string;
    notaJurado2: number;
    calificacion1: number;
    calificacion2: number;
    calificacion3: number;
    calificacion4: number;
    calificacion5: number;
    calificacion6: number;
    comentario1: string;
    comentario2: string;
    comentario3: string;
    comentario4: string;
    comentario5: string;
    comentario6: string;
    
}